+++ src/test/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerDataSourceIT.java
@@ -50,6 +50,65 @@ public class LinkedDataControllerDataSourceIT extends BaseIT {
 				false);
 	}
 
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getWqpUtTestDistance1() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/wqp/USGS-05427880/navigate/UT/wqp?distance=1",
+				HttpStatus.OK.value(),
+				FeatureTransformer.FEATURE_COUNT_HEADER,
+				"6",
+				BaseController.MIME_TYPE_GEOJSON,
+				getCompareFile(RESULT_FOLDER_WQP, "wqp_USGS-05427880_UT_wqp_distance_1.json"),
+				true,
+				false);
+
+	}
+
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getWqpUtTestDistanceEmpty() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/wqp/USGS-05427880/navigate/UT/wqp?distance=",
+				HttpStatus.OK.value(),
+				FeatureTransformer.FEATURE_COUNT_HEADER,
+				"13",
+				BaseController.MIME_TYPE_GEOJSON,
+				getCompareFile(RESULT_FOLDER_WQP, "wqp_USGS-05427880_UT_wqp_distance_empty.json"),
+				true,
+				false);
+	}
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getWqpUtTestDistanceAboveMax() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/wqp/USGS-05427880/navigate/UT/wqp?distance=10000",
+				HttpStatus.BAD_REQUEST.value(),
+				null,
+				null,
+				null,
+				"getFeatures.distance: distance must be between 1 and 9999 kilometers",
+				false,
+				false);
+	}
+
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureWqp.xml")
+	public void getWqpUtTestDistanceBelowMin() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/wqp/USGS-05427880/navigate/UT/wqp?distance=-1",
+				HttpStatus.BAD_REQUEST.value(),
+				null,
+				null,
+				null,
+				"getFeatures.distance: distance must be between 1 and 9999 kilometers",
+				false,
+				false);
+	}
+
 	//Navigation Different Datasource Testing
 	@Test
 	@DatabaseSetup("classpath:/testData/featureHuc12pp.xml")