+++ src/test/java/gov/usgs/owi/nldi/controllers/LinkedDataControllerFlowlineIT.java
@@ -16,6 +16,7 @@ import com.github.springtestdbunit.annotation.DatabaseSetup;
 import gov.usgs.owi.nldi.BaseIT;
 import gov.usgs.owi.nldi.transform.FlowLineTransformer;
 
+
 @EnableWebMvc
 @SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT)
 @DatabaseSetup("classpath:/testData/crawlerSource.xml")
@@ -64,6 +65,52 @@ public class LinkedDataControllerFlowlineIT extends BaseIT {
 				getCompareFile(RESULT_FOLDER_HUC, "huc12pp_070900020601_DM_distance_10.json"),
 				true,
 				false);
+	}
+
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureHuc12pp.xml")
+	public void getHuc12ppDM10000TestDistanceAboveMax() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/huc12pp/070900020601/navigate/DM?distance=10000",
+				HttpStatus.BAD_REQUEST.value(),
+				null,
+				null,
+				null,
+				"getFlowlines.distance: distance must be between 1 and 9999 kilometers",
+				false,
+				false);
+	}
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureHuc12pp.xml")
+	public void getHuc12ppDM0TestDistanceBelowMin() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/huc12pp/070900020601/navigate/DM?distance=-1",
+				HttpStatus.BAD_REQUEST.value(),
+				null,
+				null,
+				null,
+				"getFlowlines.distance: distance must be between 1 and 9999 kilometers",
+				false,
+				false);
+	}
+
+
+	@Test
+	@DatabaseSetup("classpath:/testData/featureHuc12pp.xml")
+	public void getHuc12ppDMTestEmptyDistance() throws Exception {
+		assertEntity(restTemplate,
+				"/linked-data/huc12pp/070900020601/navigate/DM?distance=",
+				HttpStatus.OK.value(),
+				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
+				"51",
+				BaseController.MIME_TYPE_GEOJSON,
+				getCompareFile(RESULT_FOLDER_HUC, "huc12pp_070900020601_DM_distance_empty.json"),
+				true,
+				false);
+
+
 	}
 
 	@Test