+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 *   /basin request now returns a 404 instead of 500 error when basin is not found (WDFN-334)
 *   Updated README for consistency and clearer dev setup
 *   Updated getCharacteristicData (/{featureSource}/{featureID}/{characteristicType}) to return 404s when a comid is not found
+*   Update validation of distance to accept an empty value by switching to a regex pattern
 *   Add features collection API /linked-data/{featureSource}
 *   Updated Swagger to show distance is in kilometers